Events, purchases and tickets refactored
- separated to their own files
- some changes to logic (using Optional.map() instead of
	isPresent checks
- removed old REST controller file due to its uselessness now
